pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
description = "python based access to Thunderbird mail with nicegui frontend"
keywords = [ "nicegui", "thunderbird", "mail" ]

name = "pyThunderbird"
authors = [
  { name = "Wolfgang Fahl", email = "wf@bitplan.com" }
]
maintainers = [
  { name = "Wolfgang Fahl", email = "wf@bitplan.com" }
]

dependencies= [
	"pylodstorage>=0.4.11",
	"PyYAML>=6.0",
	# https://github.com/WolfgangFahl/nicegui_widgets
	"ngwidgets>=0.11.4",
	# ftfy 6.1.3
	"ftfy>=6.1.3",
]

readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
dynamic = ["version"]

classifiers=[
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License"
]

[tool.hatch.version]
path = "thunderbird/__init__.py"

[project.urls]
Home = "https://github.com/WolfgangFahl/pyThunderbird"
Documentation = "http://wiki.bitplan.com/index.php/PyThunderbird"
Source = "https://github.com/WolfgangFahl/pyThunderbird/blob/master/thunderbird/mail.py"
IssueTracker = "https://github.com/WolfgangFahl/pyThunderbird/issues"

[project.optional-dependencies]
test = [
  "green",
]

[tool.isort]
profile = "black"

[tool.hatch.build.targets.wheel]
packages = [
  "thunderbird",
]

[project.scripts]
tbmail = "thunderbird.mail_cmd:main"
ghget = "thunderbird.github:main"
